requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
google-generativeai==0.8.3
chromadb==0.5.18
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
httpx==0.27.2
